ADK_Pipelines/pyproject.toml
[project]
name = "autostack-adk"
version = "2.0.0"
description = "AutoStack — ADK Multi-Agent Pipeline System (9 pipelines)"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.0.0",
    "uvicorn>=0.34.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
